feat: fixing bug rerender homepage table, fix bug loading on error transaction stats, change logo and favicon
