+++ src/app/page.tsx
@@ -31,6 +31,7 @@ const Home = () => {
   return (
     <>
       <Hero />
+
       <RecentTransactions />
     </>
   );
