+++ src/components/features/landing/transaction-tables/TransactionsTable.tsx
@@ -1,0 +1,19 @@
+'use client';
+
+import { DataTable } from '@/components/ui/composites/data-table/DataTable';
+
+import { useTransactionsTable } from './hook';
+
+export const TransactionsTable = () => {
+  const { getStatus, table, error } = useTransactionsTable();
+
+  return (
+    <DataTable
+      table={table}
+      fetch={{
+        status: getStatus(),
+        errorFallback: <p>Error : {error && error.message}</p>
+      }}
+    />
+  );
+};
